Extract a result class from the book creator and the book updater
