+++ lib/book_updater.rb
@@ -1,4 +1,5 @@
 require "book_validator"
+require "result"
 
 class BookUpdater
   def self.execute(book_id, attributes)
@@ -38,18 +39,5 @@ class BookUpdater
 
   def validation
     @validation ||= BookValidator.validate(book)
-  end
-
-  class Result
-    attr_reader :book, :messages
-
-    def initialize(book, messages)
-      @book = book
-      @messages = messages
-    end
-
-    def successful?
-      messages.empty?
-    end
   end
 end
